[FIX] Network 확장함수 적용
